+++ lib/geo_pattern/version.rb
@@ -1,0 +1,3 @@
+module GeoPattern
+  VERSION = "0.0.1"
+end
